fix: refacto pytest plugins
