+++ tests/decorators/conftest.py
@@ -7,6 +7,12 @@ from tests.decorators.fixtures.pipeline_fixtures import (
 )
 import pytest
 
+pytest_plugins = [
+    "tests.decorators.fixtures.pipeline_fixtures",
+    "tests.decorators.fixtures.step_fixtures",
+    "tests.fixtures.log_fixtures",
+]
+
 
 @pytest.fixture(autouse=True)
 def cleanup(temp_log_dir):
