+++ tests/models/steps/data_extraction/training/conftest.py
@@ -1,0 +1,5 @@
+pytest_plugins = [
+    "tests.fixtures.dataset_version_fixtures",
+    "tests.fixtures.initialize_integration_tests_fixtures",
+    "tests.models.steps.data_extraction.training.fixtures.training_dataset_collection_extractor_fixtures",
+]
