+++ tests/models/steps/data_validation/common/conftest.py
@@ -1,0 +1,6 @@
+pytest_plugins = [
+    "tests.fixtures.dataset_version_fixtures",
+    "tests.fixtures.initialize_integration_tests_fixtures",
+    "tests.models.dataset.common.fixtures.dataset_context_fixtures",
+    "tests.models.steps.data_validation.common.fixtures.dataset_validator_fixtures",
+]
