+++ tests/steps/data_preparation/common/conftest.py
@@ -1,0 +1,5 @@
+pytest_plugins = [
+    "tests.fixtures.dataset_version_fixtures",
+    "tests.fixtures.initialize_integration_tests_fixtures",
+    "tests.models.dataset.common.fixtures.dataset_collection_fixtures",
+]
